# optimize coin_store add_coin_records() by pipelining inserts