This give a slight speed-up when adding a new block. The sync test runs at about 7% of the original run-time.

sync test runtime (M1, SSD):

| DB sync | before | after | after / before |
| --- | --- | --- | --- |
| off | 109.85s | 102.36s | 0.932 |
| full | 116.16s | 108.02s | 0.930 |